moving everything but sampling backend to new app folder
